♻️ move deflate-worker to rum-core
