+++ packages/rum-core/src/index.ts
@@ -40,3 +40,4 @@ export * from './domain/privacy'
 export { SessionReplayState } from './domain/rumSessionManager'
 export type { RumPlugin } from './domain/plugins'
 export type { MouseEventOnElement } from './domain/action/listenActionEvents'
+export * from './domain/deflate'
