+++ packages/rum-core/test/index.ts
@@ -12,3 +12,4 @@ export * from './mockPageStateHistory'
 export * from './mockRumConfiguration'
 export * from './locationChangeSetup'
 export * from './mockContexts'
+export * from './mockWorker'
