+++ packages/rum/src/domain/segmentCollection/segment.spec.ts
@@ -1,12 +1,11 @@
 import type { DeflateEncoder, TimeStamp } from '@datadog/browser-core'
 import { noop, setDebugMode, isIE, DeflateEncoderStreamId } from '@datadog/browser-core'
-import type { RumConfiguration } from '@datadog/browser-rum-core'
+import { createDeflateEncoder, type RumConfiguration } from '@datadog/browser-rum-core'
 import { registerCleanupTask } from '@datadog/browser-core/test'
-import { MockWorker } from '../../../test'
+import { MockWorker } from '@datadog/browser-rum-core/test'
 import type { CreationReason, BrowserRecord, SegmentContext, BrowserSegment, BrowserSegmentMetadata } from '../../types'
 import { RecordType } from '../../types'
 import { getReplayStats, resetReplayStats } from '../replayStats'
-import { createDeflateEncoder } from '../deflate'
 import type { AddRecordCallback, FlushCallback, Segment } from './segment'
 import { createSegment } from './segment'
