+++ packages/rum/src/domain/segmentCollection/segmentCollection.spec.ts
@@ -1,14 +1,13 @@
 import type { ClocksState, HttpRequest, TimeStamp } from '@datadog/browser-core'
 import { DeflateEncoderStreamId, PageExitReason, isIE } from '@datadog/browser-core'
 import type { ViewHistory, ViewHistoryEntry, RumConfiguration } from '@datadog/browser-rum-core'
-import { LifeCycle, LifeCycleEventType } from '@datadog/browser-rum-core'
+import { createDeflateEncoder, LifeCycle, LifeCycleEventType } from '@datadog/browser-rum-core'
 import type { Clock } from '@datadog/browser-core/test'
 import { mockClock, registerCleanupTask, restorePageVisibility } from '@datadog/browser-core/test'
-import { createRumSessionManagerMock } from '../../../../rum-core/test'
+import { createRumSessionManagerMock, MockWorker } from '../../../../rum-core/test'
 import type { BrowserRecord, SegmentContext } from '../../types'
 import { RecordType } from '../../types'
-import { MockWorker, readMetadataFromReplayPayload } from '../../../test'
-import { createDeflateEncoder } from '../deflate'
+import { readMetadataFromReplayPayload } from '../../../test'
 import {
   computeSegmentContext,
   doStartSegmentCollection,
