+++ packages/rum/src/entries/main.ts
@@ -1,11 +1,10 @@
 // Keep the following in sync with packages/rum-slim/src/entries/main.ts
 import { defineGlobal, getGlobalObject } from '@datadog/browser-core'
 import type { RumPublicApi } from '@datadog/browser-rum-core'
-import { makeRumPublicApi, startRum } from '@datadog/browser-rum-core'
+import { makeRumPublicApi, startRum, startDeflateWorker, createDeflateEncoder } from '@datadog/browser-rum-core'
 
 import { startRecording } from '../boot/startRecording'
 import { makeRecorderApi } from '../boot/recorderApi'
-import { createDeflateEncoder, startDeflateWorker } from '../domain/deflate'
 
 export {
   CommonProperties,
